:sparkles: Feat : Item, CartItem 개발.
1. Feat: Cart Repository 추가
2. Feat: CartItem 엔티티 추가
: 상품과 장바구니는 n:m 관계이므로, 이를 해소할 장바구니상품(CartItem)이 필요함.
3. Feat: CartController, CartItemController 추가.

4. Item 속성 Count int로 변경.
: 기본 타입으로 변환하여 계산하는 로직이 필요해서 (Long -> item으로 변경.)
5. 사소한 버그 수정

+++ src/main/java/com/example/codingmall/Item/ItemDto.java
@@ -20,8 +20,8 @@ public class ItemDto {
     private String productName;
     private ItemStatus status;
 
-    private long  stock;
-    private long price;
+    private int  stock;
+    private int price;
 
     private String intro;
     private String content;
